feat: refactor final-gate to use runtime kernel and decouple cli high-level runtime
